Minor power tuning

+++ arduino/rfglow_esp32/src/main.h
@@ -4,4 +4,4 @@
 
 void setup();
 void loop();
-int vbatRead();
+float vbatRead();
